% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{distort_gt}
\alias{distort_gt}
\title{Distort the ground truth data}
\usage{
distort_gt(x, tpr = 1, tnr = 1, Q = NULL)
}
\arguments{
\item{x}{a logical vector}

\item{tpr}{true positive rate}

\item{tnr}{true negative rate}

\item{Q}{coder agreement}
}
\description{
Distort the ground truth data
}
\examples{
mod <- gen_acc(.9, .3)
ground_truth <- gen_x(0.3, size = 100)
table(ground_truth, distort_gt(ground_truth, tp = mod$tp, tn = mod$tn))  
}
