% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{gen_acc}
\alias{gen_acc}
\title{generate a random combination of tp and tn, given f1. The combination generated is always >= 0 and <= 1.}
\usage{
gen_acc(acc, pp, tp = runif(1))
}
\description{
generate a random combination of tp and tn, given f1. The combination generated is always >= 0 and <= 1.
}
