% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{gen_f1}
\alias{gen_f1}
\title{generate a random combination of tp and tn, given f1. The combination generated is always >= 0 and <= 1.}
\usage{
gen_f1(f1, pp, tp = runif(1))
}
\description{
generate a random combination of tp and tn, given f1. The combination generated is always >= 0 and <= 1.
}
