% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{simulate_pred}
\alias{simulate_pred}
\title{Stimulate the results given the predicate function}
\usage{
simulate_pred(z, fun, x, Q = 0.7, tpr = NULL, tnr = NULL)
}
\arguments{
\item{z}{an accuarcy metric}

\item{fun}{a function that generate TP and TN}

\item{x}{a data.frame of generated dataset}

\item{Q}{coder agreement}

\item{tpr}{true positive rate}

\item{tnr}{true negatve rate}
}
\description{
Stimulate the results given the predicate function
}
